requirements.txt
Package                      Version
---------------------------- --------
annotated-types              0.6.0
cachetools                   5.3.3
certifi                      2024.2.2
charset-normalizer           3.3.2
click                        8.1.7
google-ai-generativelanguage 0.6.2
google-api-core              2.18.0
google-api-python-client     2.126.0
google-auth                  2.29.0
google-auth-httplib2         0.2.0
google-generativeai          0.5.2
googleapis-common-protos     1.63.0
grpcio                       1.62.2
grpcio-status                1.62.2
httplib2                     0.22.0
idna                         3.7
pillow                       10.3.0
pillow-avif-plugin           1.4.3
pillow_heif                  0.16.0
pip                          24.0
poppler-utils                0.1.0
proto-plus                   1.23.0
protobuf                     4.25.3
pyasn1                       0.6.0
pyasn1_modules               0.4.0
pydantic                     2.7.0
pydantic_core                2.18.1
pyparsing                    3.1.2
PyPDF2                       3.0.1
python-dotenv                1.0.1
requests                     2.31.0
rsa                          4.9
setuptools                   65.5.0
tqdm                         4.66.2
typing_extensions            4.11.0
uritemplate                  4.1.1
urllib3                      2.2.1
